requirements.txt
discord.py>=2.0.0
Flask>=2.0.0
python-dotenv>=0.19.0
pytz>=2021.3
mysql-connector-python>=8.0.0